Add CV + templates

+++ Front/Template/Infos.php
@@ -1,8 +1,10 @@
 <?php
+include "../../public/user.php";
 include "../../public/hobbies.php";
 include "../../public/parcours.php";
 include "../../public/experiences.php";
 
+session_start();
 $conn = new PDO('mysql:host=localhost;dbname=database', 'root', '');
 ?>
 
@@ -15,7 +17,6 @@ $conn = new PDO('mysql:host=localhost;dbname=database', 'root', '');
 </head>
 <body>
 <?php
-session_start();
 if (isset($_SESSION['password']))  {
 ?>
 <div class="container">
@@ -44,31 +45,32 @@ if (isset($_SESSION['password']))  {
         <input type="submit" name="submit" value="Ajouter">
     </form>
     <?php
-
     $academics = $conn->query('SELECT * FROM parcours_academiques');
 
     while ($row = $academics->fetch(PDO::FETCH_ASSOC)) {
-        echo "<div class='academic-item'>";
-        echo "<p>Diplôme: " . $row['diplome'] . "</p>";
-        echo "<p>Etablissement: " . $row['etablissement'] . "</p>";
-        echo "<p>Date: " . $row['date_debut'] . " au " . $row['date_fin'] . "</p>";
-        echo "<p>Description: " . $row['description'] . "</p>";
-        echo "<button name='modif'>Modifier</button>";
-        echo "<button name='Supp'>Supprimer</button>";
-        if (isset($_POST['modif'])) {
-            echo "<form action='infos2.php' method='post'>";
-            echo "<input type='text' name='diplome' value='" . $row['diplome'] . "'>";
-            echo "<input type='text' name='etablissement' value='" . $row['etablissement'] . "'>";
-            echo "<input type='date' name='date_debut' value='" . $row['date_debut'] . "'>";
-            echo "<input type='date' name='date_fin' value='" . $row['date_fin'] . "'>";
-            echo "<input type='text' name='description' value='" . $row['description'] . "'>";
-            echo "<input type='submit' name='submit' value='Modifier'>";
-            echo "</form>";
+        if (intval(getUserID()) == $row['user_id']) {
+            echo "<div class='academic-item'>";
+            echo "<p>Diplôme: " . $row['diplome'] . "</p>";
+            echo "<p>Etablissement: " . $row['etablissement'] . "</p>";
+            echo "<p>Date: " . $row['date_debut'] . " au " . $row['date_fin'] . "</p>";
+            echo "<p>Description: " . $row['description'] . "</p>";
+            echo "<button name='modif'>Modifier</button>";
+            echo "<button name='Supp'>Supprimer</button>";
+            if (isset($_POST['modif'])) {
+                echo "<form action='infos2.php' method='post'>";
+                echo "<input type='text' name='diplome' value='" . $row['diplome'] . "'>";
+                echo "<input type='text' name='etablissement' value='" . $row['etablissement'] . "'>";
+                echo "<input type='date' name='date_debut' value='" . $row['date_debut'] . "'>";
+                echo "<input type='date' name='date_fin' value='" . $row['date_fin'] . "'>";
+                echo "<input type='text' name='description' value='" . $row['description'] . "'>";
+                echo "<input type='submit' name='submit' value='Modifier'>";
+                echo "</form>";
+            }
+            if (isset($_POST['Supp'])) {
+                deleteParcours($row['ID']);
+            }
+            echo "</div>";
         }
-        if (isset($_POST['Supp'])) {
-            deleteParcours($row['ID']);
-        }
-        echo "</div>";
     }
     ?>
 </div>
@@ -100,14 +102,29 @@ if (isset($_SESSION['password']))  {
     $experiences = $conn->query('SELECT * FROM experiences_professionnelles');
 
     while ($row = $experiences->fetch(PDO::FETCH_ASSOC)) {
-        echo "<div class='experience-item'>";
-        echo "<p>Nom de l'entreprise: " . $row['nom_entreprise'] . "</p>";
-        echo "<p>Poste: " . $row['poste'] . "</p>";
-        echo "<p>Durée: " . $row['date_debut'] . " au " . $row['date_fin'] . "</p>";
-        echo "<p>Description: " . $row['description'] . "</p>";
-        echo "<button onclick='editExperience(" . $row['ID'] . ")'>Modifier</button>";
-        echo "<button onclick='deleteExperience(" . $row['ID'] . ")'>Supprimer</button>";
-        echo "</div>";
+        if (intval(getUserID()) == $row['user_id']) {
+            echo "<div class='experience-item'>";
+            echo "<p>Nom de l'entreprise: " . $row['nom_entreprise'] . "</p>";
+            echo "<p>Poste: " . $row['poste'] . "</p>";
+            echo "<p>Durée: " . $row['date_debut'] . " au " . $row['date_fin'] . "</p>";
+            echo "<p>Description: " . $row['description'] . "</p>";
+            echo "<button name='modif'>Modifier</button>";
+            echo "<button name='Supp'>Supprimer</button>";
+            if (isset($_POST['modif'])) {
+                echo "<form action='infos2.php' method='post'>";
+                echo "<input type='text' name='nom_entreprise' value='" . $row['nom_entreprise'] . "'>";
+                echo "<input type='text' name='poste' value='" . $row['poste'] . "'>";
+                echo "<input type='date' name='date_debut' value='" . $row['date_debut'] . "'>";
+                echo "<input type='date' name='date_fin' value='" . $row['date_fin'] . "'>";
+                echo "<input type='text' name='description' value='" . $row['description'] . "'>";
+                echo "<input type='submit' name='submit' value='Modifier'>";
+                echo "</form>";
+            }
+            if (isset($_POST['Supp'])) {
+                deleteExperience($row['ID']);
+            }
+            echo "</div>";
+        }
     }
     ?>
 </div>
@@ -126,11 +143,22 @@ if (isset($_SESSION['password']))  {
     $hobbies = $conn->query('SELECT * FROM hobbies');
 
     while ($row = $hobbies->fetch(PDO::FETCH_ASSOC)) {
-        echo "<div class='hobby-item'>";
-        echo "<p>Nom du loisir: " . $row['hobby'] . "</p>";
-        echo "<button onclick='editHobby(" . $row['ID'] . ")'>Modifier</button>";
-        echo "<button onclick='deleteHobby(" . $row['ID'] . ")'>Supprimer</button>";
-        echo "</div>";
+        if (intval(getUserID()) == $row['user_id']) {
+            echo "<div class='hobby-item'>";
+            echo "<p>Nom du loisir: " . $row['hobby'] . "</p>";
+            echo "<button name='modif'>Modifier</button>";
+            echo "<button name='Supp'>Supprimer</button>";
+            if (isset($_POST['modif'])) {
+                echo "<form action='infos2.php' method='post'>";
+                echo "<input type='text' name='hobby' value='" . $row['hobby'] . "'>";
+                echo "<input type='submit' name='submit' value='Modifier'>";
+                echo "</form>";
+            }
+            if (isset($_POST['Supp'])) {
+                deleteHobby($row['ID']);
+            }
+            echo "</div>";
+        }
     }
     ?>
 </div>
